Update, Norm KO

+++ client.c
@@ -10,12 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "minitalk.h"
+#include "minitalk.h"
 
-void	send_signal(int pid, unsigned char chr)
+void send_signal(int pid, unsigned char chr)
 {
-	int				i;
-	unsigned char	temp_char;
+	int i;
+	unsigned char temp_char;
 
 	i = 8;
 	temp_char = chr;
@@ -31,21 +31,26 @@ void	send_signal(int pid, unsigned char chr)
 	}
 }
 
-int	main(int argc, char **argv)
+int main(int argc, char **argv)
 {
-	int		server_pid;
-	const char	*msg;
-	int			i;
+	int server_pid;
+	const char *msg;
+	int i;
 
 	if (argc == 3)
-    {
-	    server_pid = ft_atoi(argv[1]);
-	    msg = argv[2];
-	    i = 0;
-	    while (msg[i])
-	    	send_signal(server_pid, msg[i++]);
-	    send_signal(server_pid, '\0');
-    }
-    else
-        ft_printf("ERROR!!!!\n");
+	{
+		server_pid = ft_atoi(argv[1]);
+		if (server_pid < 0)
+		{
+			ft_printf("ERROR!!!!\n");
+			return (0);
+		}
+		msg = argv[2];
+		i = 0;
+		while (msg[i])
+			send_signal(server_pid, msg[i++]);
+		send_signal(server_pid, '\0');
+	}
+	else
+		ft_printf("ERROR!!!!\n");
 }
